+++ packages/polyserve/CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 <!-- ## Unreleased -->
 <!-- Add new, unreleased items here. -->
 
+## [0.27.6] (2018-04-18)
+* Fix node module resolution for the case where the root package is served from
+  the components/ directory and imports a module from its own package using a
+  path.
+
 ## [0.27.5] (2018-04-17)
 * Update --npm flag description.